Move group msg encryption primitive into session encrypt, avoids circular dependency
