+++ src/config/internal.cpp
@@ -5,7 +5,6 @@
 #include <oxenc/base64.h>
 #include <oxenc/bt_value_producer.h>
 #include <oxenc/hex.h>
-#include <zstd.h>
 
 #include <iterator>
 #include <optional>
@@ -242,62 +241,4 @@ void load_unknowns(
             throw oxenc::bt_deserialize_invalid{"invalid bencoded value type"};
     }
 }
-
-namespace {
-    struct zstd_decomp_freer {
-        void operator()(ZSTD_DStream* z) const { ZSTD_freeDStream(z); }
-    };
-
-    using zstd_decomp_ptr = std::unique_ptr<ZSTD_DStream, zstd_decomp_freer>;
-}  // namespace
-
-std::vector<unsigned char> zstd_compress(
-        std::span<const unsigned char> data, int level, std::span<const unsigned char> prefix) {
-    std::vector<unsigned char> compressed;
-    if (prefix.empty())
-        compressed.resize(ZSTD_compressBound(data.size()));
-    else {
-        compressed.resize(prefix.size() + ZSTD_compressBound(data.size()));
-        std::copy(prefix.begin(), prefix.end(), compressed.begin());
-    }
-    auto size = ZSTD_compress(
-            compressed.data() + prefix.size(),
-            compressed.size() - prefix.size(),
-            data.data(),
-            data.size(),
-            level);
-    if (ZSTD_isError(size))
-        throw std::runtime_error{"Compression failed: " + std::string{ZSTD_getErrorName(size)}};
-
-    compressed.resize(prefix.size() + size);
-    return compressed;
-}
-
-std::optional<std::vector<unsigned char>> zstd_decompress(
-        std::span<const unsigned char> data, size_t max_size) {
-    zstd_decomp_ptr z_decompressor{ZSTD_createDStream()};
-    auto* zds = z_decompressor.get();
-
-    ZSTD_initDStream(zds);
-    ZSTD_inBuffer input{/*.src=*/data.data(), /*.size=*/data.size(), /*.pos=*/0};
-    std::array<unsigned char, 4096> out_buf;
-    ZSTD_outBuffer output{/*.dst=*/out_buf.data(), /*.size=*/out_buf.size(), /*.pos=*/0};
-
-    std::vector<unsigned char> decompressed;
-
-    size_t ret;
-    do {
-        output.pos = 0;
-        if (ret = ZSTD_decompressStream(zds, &output, &input); ZSTD_isError(ret))
-            return std::nullopt;
-
-        if (max_size > 0 && decompressed.size() + output.pos > max_size)
-            return std::nullopt;
-
-        decompressed.insert(decompressed.end(), out_buf.begin(), out_buf.begin() + output.pos);
-    } while (ret > 0 || input.pos < input.size);
-
-    return decompressed;
-}
-
 }  // namespace session::config
